Mapping Vive to Hololens world-space
Using a registration point at 0,0,0:
Line up the center of the physical controller with the registration point.
Tap the menu button on the physical controller.
The virtual controller should now be mapped to the physical controller.

Unfortunately the Coordinate spaces btw Vive and Hololens may be rotated
so the further you get from the origin, the more the physical
and virtual controllers drift apart.